pyproject.toml
[project]
name = "semanticng"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "pydantic>=2.6",
  "pydantic-settings>=2.2",
  "orjson>=3.10",
  "httpx>=0.27",
  "tenacity>=9.0",
  "rich>=13.7",
]

[project.optional-dependencies]
test = [
  "pytest>=8",
  "pytest-cov>=5",
  "mypy>=1.9",
  "ruff>=0.6",
  "behave>=1.2.6",
]
gherkin = [
  "gherkin-official>=30",          # if you generate/parse .feature AST
  "cucumber-expressions>=19.0.0",  # if you want cucumber-expr matching
]
es = ["elasticsearch>=8.15"]
ha = ["HomeAssistant-API>=5.0.2.post2"]
ollama = ["ollama>=0.6"]
lang = [
  "langgraph>=1.0",
  "langchain-core>=0.3",  # only if truly used
]


[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_return_any = true
no_implicit_optional = true
disallow_any_generics = true
disallow_subclassing_any = true
check_untyped_defs = true

# If your project imports optional deps in some modules:
ignore_missing_imports = false

# Tests often use duck-typing/fakes:
[[tool.mypy.overrides]]
module = "tests.*"
disallow_any_generics = false
check_untyped_defs = false
strict = false
